fix: Refactor gold to Long and force prod profile

+++ backend/src/main/java/com/tycoon/forge/domain/user/entity/User.java
@@ -4,7 +4,6 @@ import com.tycoon.forge.domain.common.BaseTimeEntity;
 import jakarta.persistence.*;
 import lombok.*;
 
-import java.math.BigInteger;
 import java.util.UUID;
 
 @Getter
@@ -25,7 +24,7 @@ public class User extends BaseTimeEntity {
 
     @Builder.Default
     @Column(nullable = false)
-    private BigInteger gold = BigInteger.valueOf(10000);
+    private Long gold = 10000L;
 
     @Builder.Default
     @Column(nullable = false)
